improve the way to get port and ip
